Repository input interface

+++ include/App/TransportParser/Read.hpp
@@ -29,7 +29,7 @@ namespace msg{
     };
 
     std::string nodeId;
-    uint8_t depth;
+    uint16_t depth;
     Class nodeType;
     QueryMode attribsReq;
     std::vector<Feature> features;
